+++ e2e/flows/binary-files.e2e.2.js
@@ -5,6 +5,7 @@ import chai, { expect } from 'chai';
 import Helper from '../e2e-helper';
 import { statusWorkspaceIsCleanMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 import NpmCiRegistry, { supportNpmCiRegistryTesting } from '../npm-ci-registry';
+import { AUTO_GENERATED_STAMP } from '../../src/constants';
 
 chai.use(require('chai-fs'));
 
@@ -487,6 +488,24 @@ describe('binary files', function () {
           );
         });
       });
+    });
+  });
+  describe('export an md file with the same name as another js file with compiler (bug #1628)', () => {
+    before(() => {
+      helper.setNewLocalAndRemoteScopes();
+      helper.createFile('bar', 'my-comp.md', 'some md5 content');
+      helper.createFile('bar', 'my-comp.js');
+      helper.runCmd('bit add bar -m my-comp.js -i bar/foo');
+      helper.tagAllComponents();
+      helper.exportAllComponents();
+      helper.importDummyCompiler();
+      helper.tagAllComponents();
+      helper.exportAllComponents();
+    });
+    it('should not overwrite the md file with an auto-generated content', () => {
+      const mdContent = helper.readFile('bar/my-comp.md');
+      expect(mdContent).to.not.have.string(AUTO_GENERATED_STAMP);
+      expect(mdContent).to.have.string('some md5 content');
     });
   });
 });