+++ src/consumer/component-ops/many-components-writer.js
@@ -6,7 +6,7 @@ import pMapSeries from 'p-map-series';
 import { moveExistingComponent } from './move-components';
 import { getAllComponentsLinks } from '../../links';
 import { installNpmPackagesForComponents } from '../../npm-client/install-packages';
-import * as packageJson from '../component/package-json';
+import * as packageJsonUtils from '../component/package-json-utils';
 import type { ComponentWithDependencies } from '../../scope';
 import type Component from '../component/consumer-component';
 import { COMPONENT_ORIGINS } from '../../constants';
@@ -110,9 +110,9 @@ export default class ManyComponentsWriter {
   }
   async _installPackages() {
     logger.debug('ManyComponentsWriter, _installPackages');
-    await packageJson.addWorkspacesToPackageJson(this.consumer, this.writeToPath);
+    await packageJsonUtils.addWorkspacesToPackageJson(this.consumer, this.writeToPath);
     if (this.addToRootPackageJson) {
-      await packageJson.addComponentsToRoot(this.consumer, this.writtenComponents);
+      await packageJsonUtils.addComponentsToRoot(this.consumer, this.writtenComponents);
     }
     await this._installPackagesIfNeeded();
   }
@@ -231,21 +231,6 @@ export default class ManyComponentsWriter {
             'writeToComponentsDir, ignore dependency {dependencyId} as it already exists in bit map',
             { dependencyId }
           );
-          this.consumer.bitMap.addDependencyToParent(componentWithDeps.component.id, dependencyId);
-          return Promise.resolve(dep);
-        }
-        if (
-          depFromBitMap &&
-          (fs.existsSync(depFromBitMap.rootDir) ||
-            this.writtenComponents.find(c => c.writtenPath === depFromBitMap.rootDir))
-        ) {
-          dep.writtenPath = depFromBitMap.rootDir;
-          logger.debugAndAddBreadCrumb(
-            'writeToComponentsDir',
-            'writeToComponentsDir, ignore dependency {dependencyId} as it already exists in bit map and file system',
-            { dependencyId }
-          );
-          this.consumer.bitMap.addDependencyToParent(componentWithDeps.component.id, dependencyId);
           return Promise.resolve(dep);
         }
         if (this.dependenciesIdsCache[dependencyId]) {
@@ -255,7 +240,6 @@ export default class ManyComponentsWriter {
             { dependencyId }
           );
           dep.writtenPath = this.dependenciesIdsCache[dependencyId];
-          this.consumer.bitMap.addDependencyToParent(componentWithDeps.component.id, dependencyId);
           return Promise.resolve(dep);
         }
         const depRootPath = this.consumer.composeRelativeDependencyPath(dep.id);
@@ -270,7 +254,6 @@ export default class ManyComponentsWriter {
           component: dep,
           writeToPath: depRootPath,
           origin: COMPONENT_ORIGINS.NESTED,
-          parent: componentWithDeps.component.id,
           existingComponentMap: componentMap
         });
         return componentWriter.populateComponentsFilesToWrite();