+++ src/links/node-modules-linker.js
@@ -4,12 +4,12 @@ import R from 'ramda';
 import glob from 'glob';
 import { BitId } from '../bit-id';
 import type Component from '../consumer/component/consumer-component';
-import { COMPONENT_ORIGINS } from '../constants';
+import { COMPONENT_ORIGINS, PACKAGE_JSON } from '../constants';
 import type ComponentMap from '../consumer/bit-map/component-map';
 import logger from '../logger/logger';
 import { pathRelativeLinux, first, pathNormalizeToLinux } from '../utils';
 import type Consumer from '../consumer/consumer';
-import { getIndexFileName, getComponentsDependenciesLinks } from './link-generator';
+import { getComponentsDependenciesLinks } from './link-generator';
 import { getLinkToFileContent } from './link-content';
 import type { PathOsBasedRelative, PathLinuxRelative } from '../utils/path';
 import getNodeModulesPathOfComponent from '../utils/bit/component-node-modules-path';
@@ -19,7 +19,7 @@ import Symlink from './symlink';
 import DataToPersist from '../consumer/component/sources/data-to-persist';
 import LinkFile from './link-file';
 import ComponentsList from '../consumer/component/components-list';
-import { preparePackageJsonToWrite, addPackageJsonDataToPersist } from '../consumer/component/package-json';
+import { preparePackageJsonToWrite } from '../consumer/component/package-json-utils';
 
 type LinkDetail = { from: string, to: string };
 export type LinksResult = {
@@ -117,7 +117,9 @@ export default class NodeModuleLinker {
       const distTarget = component.dists.getDistDir(this.consumer, componentMap.getRootDir());
       const packagesSymlinks = this._getSymlinkPackages(srcTarget, distTarget, component);
       this.dataToPersist.addManySymlinks(packagesSymlinks);
-      this.dataToPersist.addSymlink(Symlink.makeInstance(distTarget, linkPath, componentId));
+      const distSymlink = Symlink.makeInstance(distTarget, linkPath, componentId);
+      distSymlink.forDistOutsideComponentsDir = true;
+      this.dataToPersist.addSymlink(distSymlink);
     } else {
       this.dataToPersist.addSymlink(Symlink.makeInstance(srcTarget, linkPath, componentId));
     }
@@ -222,7 +224,9 @@ export default class NodeModuleLinker {
         // dir into the dist dir. (see consumer-component.write())
         const from = component.dists.getDistDirForConsumer(this.consumer, parentRootDir);
         const to = component.dists.getDistDirForConsumer(this.consumer, dependencyRootDir);
-        dependenciesLinks.push(this._getDependencyLink(from, dependency.id, to, component.bindingPrefix));
+        const distSymlink = this._getDependencyLink(from, dependency.id, to, component.bindingPrefix);
+        distSymlink.forDistOutsideComponentsDir = true;
+        dependenciesLinks.push(distSymlink);
       }
       return dependenciesLinks;
     };
@@ -292,9 +296,11 @@ export default class NodeModuleLinker {
    * It makes it easier for Author to use absolute syntax between their own components.
    */
   _createPackageJsonForAuthor(component: Component) {
+    const hasPackageJsonAsComponentFile = component.files.some(file => file.relative === PACKAGE_JSON);
+    if (hasPackageJsonAsComponentFile) return; // don't generate package.json on top of the user package.json
     const dest = path.join(getNodeModulesPathOfComponent(component.bindingPrefix, component.id));
     const { packageJson } = preparePackageJsonToWrite(this.consumer, component, dest, true);
-    addPackageJsonDataToPersist(packageJson, this.dataToPersist);
+    this.dataToPersist.addFile(packageJson.toJSONFile());
   }
 
   /**